% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{api_status}
\alias{api_status}
\title{Check API endpoint health}
\usage{
api_status()
}
\value{
A tibble with endpoint status information
}
\description{
Test connectivity to realtor.com API endpoints.
}
\examples{
\dontrun{
api_status()
}
}
