% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_info}
\alias{cache_info}
\title{Get cache information}
\usage{
cache_info()
}
\value{
A list with cache information
}
\description{
Returns information about the realtR cache including size and location.
}
\examples{
\dontrun{
cache_info()
}
}
