% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{dictionary_listing_features}
\alias{dictionary_listing_features}
\title{Property feature dictionary}
\usage{
dictionary_listing_features()
}
\value{
a \code{tibble}
}
\description{
This function returns a dictionary
of searchable property features.  These features
can be used as search parameters in the \link{listing_counts}, \link{listings}, \link{listing_counts}
and \link{map_listings} functions.
}
\examples{
dictionary_listing_features()
}
\seealso{
Other dictionary: \code{\link{dictionary_property_types}}

Other listing search: \code{\link{dictionary_property_types}},
  \code{\link{listing_counts}}, \code{\link{listings}},
  \code{\link{map_listings}}
}
\concept{dictionary}
\concept{listing search}
