% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{dictionary_property_types}
\alias{dictionary_property_types}
\title{Property type dictionary}
\usage{
dictionary_property_types()
}
\value{
a \code{data_frame}
}
\description{
Searchable property types
for isolating listings
}
\examples{
dictionary_property_types()
}
\seealso{
Other dictionary: \code{\link{dictionary_listing_features}}

Other listing search: \code{\link{dictionary_listing_features}},
  \code{\link{listing_counts}}, \code{\link{listings}},
  \code{\link{map_listings}}
}
\concept{dictionary}
\concept{listing search}
