% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{geocode}
\alias{geocode}
\title{Location geocoder}
\usage{
geocode(locations = NULL, search_types = c("neighborhood", "city", "county",
  "postal_code", "address", "building", "street", "school"), limit = 100,
  return_message = TRUE, ...)
}
\arguments{
\item{locations}{vector of locations}

\item{search_types}{vector of search parameters options include \itemize{
\item neighborhood - includes neighborhood information
\item city - includes city information
\item county - includes county information
\item postal_code - includes zipcode
\item building - include building info
\item street - include street info
\item school - include school info
}}

\item{limit}{numeric vector of results cannot exceed 100}

\item{return_message}{if \code{TRUE} returns a message}

\item{...}{extra parameters}
}
\value{
a \code{data_frame}
}
\description{
This function geocodes a users vector of locations
and returns a \code{data_frame} with the corresponding results
}
\examples{
geocode(locations = c("Palm Springs", "Bethesda", 10016), limit = 100)
}
\concept{geocoder}
