% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{get_aws_geo_areas}
\alias{get_aws_geo_areas}
\title{Get AWS area geo-query}
\usage{
get_aws_geo_areas(area_names = NULL, search_types = c("neighborhood",
  "city", "county", "postal_code", "address", "building", "street", "school"),
  limit = 100, return_message = TRUE, ...)
}
\arguments{
\item{area_names}{vector of area names}

\item{search_types}{vector of search parameters options include \itemize{
\item neighborhood - includes neighborhood information
\item city - includes city information
\item county - includes county information
\item postal_code - includes zipcode
\item building - include building info
\item street - include street info
\item school - include school info
}}

\item{limit}{numeric vector of results cannot exceed 100}

\item{return_message}{if \code{TRUE} returns a message}

\item{...}{}
}
\description{
This function returns data from the Amazon geoquery API
for the specified areas
}
\examples{
get_aws_geo_areas(area_names = c("Palm Springs", "Bethesda"), limit = 100)
}
