% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtor.R
\name{get_markets_trends}
\alias{get_markets_trends}
\title{Acquires market trend data}
\usage{
get_markets_trends(location_names = NULL, return_wide = TRUE,
  return_message = F, ...)
}
\arguments{
\item{location_names}{vector of location names, location name must contain
a city name and a comma ie "Brooklyn, NY" or a zipcode}

\item{return_wide}{if \code{TRUE} returns data in wide form}

\item{return_message}{if \code{TRUE} returns a message}

\item{...}{extra parameters}
}
\value{
a \code{data_frame}
}
\description{
This function returns market
trend information dating back to 2015
for the user's specified locations
}
\examples{
get_markets_trends(location_names = c("Greenwich, CT", "New London, CT", "Woodside, CA"), return_message = F, return_wide = T)
}
