% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtor.R
\name{get_mortage_rates}
\alias{get_mortage_rates}
\title{Gets current mortgage rates}
\usage{
get_mortage_rates(return_wide = F)
}
\arguments{
\item{return_wide}{if \code{TRUE} widens data and removes duration and benchmark variables}
}
\description{
Gets current mortgage rates
}
\examples{
get_mortage_rates(return_wide = F)
}
