% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtor.R
\name{get_realtor_locations_median_prices}
\alias{get_realtor_locations_median_prices}
\title{Get median market statistics for specified locations}
\usage{
get_realtor_locations_median_prices(location_names = NULL,
  return_message = TRUE, ...)
}
\arguments{
\item{location_names}{vector of location names, location name must contain
a city name and a comma ie "Brooklyn, NY"}

\item{return_message}{if \code{TRUE} returns a message}

\item{...}{}
}
\value{
a \code{data_frame}
}
\description{
Returns summary market information for the specified
location.  The locaiton name must be a city bounded by a comma
}
\examples{
get_realtor_locations_median_prices(location_names = "Greenwich, CT")
}
