% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{html_listing_urls}
\alias{html_listing_urls}
\title{Parse html listings}
\usage{
html_listing_urls(urls = NULL, include_features = F, sleep_time = 1,
  return_message = TRUE)
}
\arguments{
\item{urls}{vector of html listings}

\item{include_features}{if \code{TRUE} includes features}

\item{sleep_time}{if not \code{NULL} sleep time between scrapes}

\item{return_message}{if \code{TRUE} returns message}
}
\description{
Parse vector of listing html listing
urls.
}
\details{
This function will likely result in DDOS so use carefully.
When possible find the JSON api data for the listings.
}
\examples{
html_listing_urls(urls = "https://www.realtor.com/realestateandhomes-detail/5301-Westbard-Cir-Apt-323_Bethesda_MD_20816_M63437-59115")
}
