% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_listing_data}
\alias{join_listing_data}
\title{Join meta level listing data and detailed listing data}
\usage{
join_listing_data(data_listings, data_detailed_listings)
}
\arguments{
\item{data_listings}{\code{data_frame} from \code{listings} function}

\item{data_detailed_listings}{\code{data_frame} from \code{parse_listing_urls}}
}
\value{
a \code{data_frame}
}
\description{
Join meta level listing data and detailed listing data
}
