% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{listing_counts}
\alias{listing_counts}
\title{Title}
\usage{
listing_counts(locations, search_type = "city", city_isolated = NULL,
  county_isolated = NULL, zipcode_isolated = NULL, state_isolated = NULL,
  street_isolated = NULL, neighborhood_isolated = NULL, beds_min = NULL,
  beds_max = NULL, baths_min = NULL, baths_max = NULL, price_min = NULL,
  price_max = NULL, only_open_houses = NULL, property_type = NULL,
  sqft_min = NULL, sqft_max = NULL, acre_min = NULL, acre_max = NULL,
  age_min = NULL, age_max = NULL, days_on_market = NULL, pending = NULL,
  is_new_construction = NULL, is_foreclosure = NULL,
  include_pending_contingency = TRUE)
}
\arguments{
\item{include_pending_contingency}{}
}
\description{
Title
}
