% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{listings}
\alias{listings}
\title{Title}
\usage{
listings(locations = NULL, search_type = "city", city_isolated = NULL,
  county_isolated = NULL, zipcode_isolated = NULL, state_isolated = NULL,
  street_isolated = NULL, features = NULL, only_open_houses = NULL,
  neighborhood_isolated = NULL, beds_min = NULL, beds_max = NULL,
  baths_min = NULL, baths_max = NULL, price_min = NULL,
  price_max = NULL, property_type = NULL, sqft_min = NULL,
  sqft_max = NULL, acre_min = NULL, acre_max = NULL, age_min = NULL,
  age_max = NULL, days_on_market = NULL, pending = NULL,
  is_new_construction = NULL, is_foreclosure = NULL,
  include_pending_contingency = TRUE)
}
\arguments{
\item{locations}{}

\item{search_type}{}

\item{city_isolated}{}

\item{county_isolated}{}

\item{zipcode_isolated}{}

\item{state_isolated}{}

\item{street_isolated}{}

\item{features}{}

\item{only_open_houses}{}

\item{neighborhood_isolated}{}

\item{beds_min}{}

\item{beds_max}{}

\item{baths_min}{}

\item{baths_max}{}

\item{price_min}{}

\item{price_max}{}

\item{property_type}{}

\item{sqft_min}{}

\item{sqft_max}{}

\item{acre_min}{}

\item{acre_max}{}

\item{age_min}{}

\item{age_max}{}

\item{days_on_market}{}

\item{pending}{}

\item{is_new_construction}{}

\item{is_foreclosure}{}

\item{include_pending_contingency}{}
}
\description{
Title
}
