% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mortage_rates}
\alias{mortage_rates}
\title{Gets current mortgage rates}
\usage{
mortage_rates(return_wide = F)
}
\arguments{
\item{return_wide}{if \code{TRUE} widens data and removes duration and benchmark variables}
}
\value{
a \code{data_frame}
}
\description{
Gets current mortgage rates
}
\examples{
mortage_rates(return_wide = F)
}
