% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mortgage_rates}
\alias{mortgage_rates}
\title{Mortgage Rates (Deprecated)}
\usage{
mortgage_rates(return_wide = FALSE)
}
\arguments{
\item{return_wide}{if \code{TRUE} widens data}
}
\value{
an empty \code{tibble} with a deprecation warning
}
\description{
`r lifecycle::badge("deprecated")`

This function is deprecated. Use \code{\link{mortgage_rates_fred}} instead,
which fetches data from the Federal Reserve Economic Data (FRED) database.
}
\examples{
\dontrun{
# Use mortgage_rates_fred() instead
mortgage_rates_fred()
}
}
\seealso{
Other interest rates: 
\code{\link{mortgage_rates_fred}()}
}
\concept{interest rates}
