% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{parse_geo_urls}
\alias{parse_geo_urls}
\title{Parse GEO URLS}
\usage{
parse_geo_urls(
  urls =
    "https://parser-external.geo.moveaws.com/suggest?input=bethesda&limit=100&client_id=rdcV8&area_types=neighborhood,city,county,postal_code,address",
  .parallel = NULL,
  return_message = TRUE
)
}
\arguments{
\item{urls}{}

\item{return_message}{}

\item{use_future}{}
}
\description{
Parse GEO URLS
}
