% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing_api.R
\name{parse_listing_urls}
\alias{parse_listing_urls}
\title{Parse listing urls}
\usage{
parse_listing_urls(urls = NULL, sleep_time = 1,
  assign_to_environment = F, return_message = TRUE)
}
\arguments{
\item{urls}{vector of urls}

\item{sleep_time}{sleep time in between url}

\item{assign_to_environment}{if \code{TRUE} assign to environment}

\item{return_message}{if \code{TRUE} returns a message - default \code{FALSE}}
}
\value{
a \code{tibble}
}
\description{
Parses a vector of listing API urls
}
