% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_price.R
\name{predict_price}
\alias{predict_price}
\title{Predicts price}
\usage{
predict_price(city = "Минск", district = "Заводской р-н",
  rooms = 2, floor = 3, total_floors = 8, square = 45,
  square_useful = 28, kitchen = 6.5, type = "панельный")
}
\arguments{
\item{city}{City name}

\item{district}{District name}

\item{rooms}{Number of rooms}

\item{floor}{Floor number}

\item{total_floors}{Total floors in the house}

\item{square}{Square meters for the flat}

\item{square_useful}{Useful square}

\item{kitchen}{Square meters for kitchen}

\item{type}{House type}
}
\value{
Flat price
}
\description{
Predicts price
}

