% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{properties_near}
\alias{properties_near}
\title{Properties near a location}
\usage{
properties_near(locations = NULL, return_message = TRUE)
}
\arguments{
\item{locations}{a vector of locations}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{

}
\description{
This function returns 50
properties near a specified location.
}
\details{
The location can be an exact address, zipcode
city or a neighborhood.
}
\examples{
locations <-  c("2449 Tracy Place, NW, Washington DC", "Sunset Island, Miami Beach, FL" )
properties_near(locations = locations)
}
