% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtR-package.R
\docType{package}
\name{realtoR}
\alias{realtR}
\alias{realtR-package}
\alias{realtoR}
\title{realtR}
\description{
Tools to make finding property
suck less
}
\author{
Alex Bresler (abresler@asbcllc.com)
}
