% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtR-package.R
\docType{package}
\name{realtR-package}
\alias{realtR}
\alias{realtR-package}
\title{realtR: Real Estate Data Tools}
\description{
Tools to make finding property data suck less.
Provides programmatic access to residential real estate
information from realtor.com.
}
\author{
Alex Bresler (abresler@asbcllc.com)
}
\keyword{internal}
