% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rental.R
\name{rental_estimates}
\alias{rental_estimates}
\title{Moderate rental estimates}
\usage{
rental_estimates(zipcodes = NULL, return_message = T)
}
\arguments{
\item{zipcodes}{vector of zipcodes}

\item{return_message}{if \code{TRUE} returns a message}
}
\description{
Returns moderate rental estimates for 
specified zipcodes
}
\examples{
\dontrun{
rental_estimates(zipcodes = c("20852", "10016", "90210"), return_message = TRUE)
}
}
