% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broker_bs.R
\name{summarise_broker_bullshit}
\alias{summarise_broker_bullshit}
\title{Clean Broker Bull Shit}
\usage{
summarise_broker_bullshit(
  descriptions,
  tagger,
  sentences = 1,
  return_message = TRUE
)
}
\arguments{
\item{descriptions}{vector of descriptions}

\item{tagger}{a udpipe object - this is the english dictionary on your computer from the udpipe package}

\item{sentences}{number of sentences to whittle down to}

\item{return_message}{if \code{TRUE} returns messages}
}
\description{
This function takes a 
vector descriptions and cleans
them into a concise sentence or 
sentences
}
\examples{
\dontrun{
library(udpipe)
library(textrank)
# Download model first: udpipe_download_model(language = "english")
tagger <- udpipe_load_model(file = "path/to/english-model.udpipe")
broker_bs <- "Beautiful home with stunning views and modern amenities."
summarise_broker_bullshit(descriptions = broker_bs, tagger = tagger, sentences = 1)
}

}
