% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{table_listings}
\alias{table_listings}
\title{Property Listings tables}
\usage{
table_listings(locations, include_features = F, radius = NULL,
  parse_property_details = F, return_message = T, ...)
}
\arguments{
\item{locations}{}

\item{include_features}{}

\item{radius}{}

\item{parse_property_details}{}

\item{return_message}{}

\item{...}{}
}
\description{
Property Listings tables
}
