% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtor.R
\name{tabllistings}
\alias{tabllistings}
\title{Property Listings}
\usage{
tabllistings(locations, include_features = F, radius = NULL,
  parse_property_details = F, return_message = T, ...)
}
\arguments{
\item{locations}{}

\item{include_features}{}

\item{radius}{}

\item{parse_property_details}{}

\item{return_message}{}

\item{...}{}
}
\description{
Property Listings
}
