% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trends.R
\name{trends_zipcodes}
\alias{trends_zipcodes}
\title{Market trends for zipcodes}
\usage{
trends_zipcodes(zipcodes = NULL, return_message = T)
}
\arguments{
\item{zipcodes}{vector of zipcodes}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
\code{tibble}
}
\description{
Market trends for zipcodes
}
\examples{
\dontrun{
trends_zipcodes(zipcodes = c("90210", "10016"))
}
}
