% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{listings}
\alias{listings}
\title{MLS listing data}
\usage{
listings(locations = NULL, search_type = "city",
  city_isolated = NULL, county_isolated = NULL,
  zipcode_isolated = NULL, state_isolated = NULL,
  street_isolated = NULL, features = NULL, only_open_houses = NULL,
  neighborhood_isolated = NULL, beds_min = NULL, beds_max = NULL,
  baths_min = NULL, baths_max = NULL, price_min = NULL,
  price_max = NULL, property_type = NULL, sqft_min = NULL,
  sqft_max = NULL, acre_min = NULL, acre_max = NULL,
  age_min = NULL, age_max = NULL, days_on_market = NULL,
  pending = NULL, is_new_construction = NULL,
  generate_new_cookies = F, include_pending_contingency = TRUE,
  sleep_time = 5)
}
\arguments{
\item{locations}{vector of locations}

\item{search_type}{search type options include \itemize{
\item city - \code{default}
\item county
}}

\item{city_isolated}{if not \code{NULL} isolates}

\item{county_isolated}{if not \code{NULL} isolates county}

\item{zipcode_isolated}{if not \code{NULL} isolates zipcode}

\item{state_isolated}{if not \code{NULL} isolates state}

\item{street_isolated}{if not \code{NULL} isolates street}

\item{features}{if not \code{NULL} list of searchable features
see \link{dictionary_listing_features} for options}

\item{only_open_houses}{if \code{TRUE} isolates open houses}

\item{neighborhood_isolated}{if not \code{NULL} isolates}

\item{beds_min}{if not \code{NULL} minimum bedrooms}

\item{beds_max}{if not \code{NULL} maximum bedrooms}

\item{baths_min}{if not \code{NULL} minimum bathrooms}

\item{baths_max}{if not \code{NULL} maximum bathrooms}

\item{price_min}{if not \code{NULL} minimum price}

\item{price_max}{if not \code{NULL} maximum price}

\item{property_type}{if not \code{NULL} type of property options
see \link{dictionary_property_types} for options}

\item{sqft_min}{if not \code{NULL} minimum square footage}

\item{sqft_max}{if not \code{NULL} maximum square footage}

\item{acre_min}{if not \code{NULL} minimum acres}

\item{acre_max}{if not \code{NULL} maximum acres}

\item{age_min}{if not \code{NULL} minimum age}

\item{age_max}{if not \code{NULL} maximum age}

\item{days_on_market}{if not \code{NULL} count of days on market}

\item{pending}{if \code{TRUE} include pending}

\item{is_new_construction}{if \code{TRUE} isolates to new construction}

\item{generate_new_cookies}{generate new cookies}

\item{include_pending_contingency}{if \code{TRUE} also includes pending and contingent sales}

\item{sleep_time}{sleep time}
}
\value{
a \code{data_frame}
}
\description{
Returns MLS data for
specified locations and parameters
}
\examples{
library(realtR)
 library(dplyr)
 df_big_ass_houses_with_pools <-
 listings(
 locations = c(
 "Buckhead, Atlanta, GA",
 90210,
 "Greenwich, CT",
 "Malibu, CA",
 "Soho, New York, NY"
 ),
 beds_min = 4,
 features = 'Swimming Pool',
 sqft_min = 3000
 )
}
\seealso{
Other listing search: \code{\link{dictionary_listing_features}},
  \code{\link{dictionary_property_types}},
  \code{\link{listing_counts}}, \code{\link{map_listings}}
}
\concept{listing search}
