% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{median_prices}
\alias{median_prices}
\title{Median market statistics}
\usage{
median_prices(locations = NULL, return_message = TRUE, ...)
}
\arguments{
\item{locations}{vector of location , location name must contain
a city name and a comma ie "Brooklyn, NY"}

\item{return_message}{if \code{TRUE} returns a message}

\item{...}{extra parameters}
}
\value{
a \code{data_frame}
}
\description{
Returns summary median market information for the specified
locations.
}
\details{
The location names must be a city and/or neighborhood bounded
by commas.
}
\examples{
median_prices(locations = c("Greenwich, CT", "New London, CT", "Woodside, CA", "Park Slope, Brooklyn, NY"))
}
\concept{market information}
