% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mortage_rates}
\alias{mortage_rates}
\title{Motgage Rates}
\usage{
mortage_rates(return_wide = F)
}
\arguments{
\item{return_wide}{if \code{TRUE} widens data and removes duration and benchmark variables}
}
\value{
a \code{data_frame}
}
\description{
Returns a variety of 
interest rates for various
mortgage types.
}
\examples{
mortage_rates(return_wide = F)
}
\concept{interest rates}
