% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mortgage_rates}
\alias{mortgage_rates}
\title{Mortgage Rates}
\usage{
mortgage_rates(return_wide = F)
}
\arguments{
\item{return_wide}{if \code{TRUE} widens data and removes duration and benchmark variables}
}
\value{
a \code{data_frame}
}
\description{
Returns a variety of
interest rates for various
mortgage types.
}
\examples{
mortgage_rates(return_wide = F)
}
