% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtor.R
\name{parse_listing_urls}
\alias{parse_listing_urls}
\title{Parse listing urls}
\usage{
parse_listing_urls(urls = NULL, include_features = F,
  return_message = TRUE)
}
\arguments{
\item{urls}{vector of urls}

\item{include_features}{if \code{TRUE} includes features}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame}
}
\description{
Parse listing urls
}
