% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{table_listings}
\alias{table_listings}
\title{Property listing table}
\usage{
table_listings(locations, include_features = F, radius = NULL,
  parse_property_details = F, return_message = T, ...)
}
\arguments{
\item{locations}{vector of table listings}

\item{include_features}{if \code{TRUE} includes property featues from pages}

\item{radius}{if not \code{NULL} additional search radius}

\item{parse_property_details}{if \code{TRUE}}

\item{return_message}{if \code{TRUE} returns messages}

\item{...}{extra parameters}
}
\value{
a \code{data_frame}
}
\description{
Returns listing table.  Slower and not advised 
unless \link{listings()} does not work for your search.
}
\examples{
table_listings(locations = "Marietta, GA")
}
