% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{vitality}
\alias{vitality}
\title{Market vitality}
\usage{
vitality(locations = NULL, return_message = TRUE, ...)
}
\arguments{
\item{locations}{vector of location names, location name must contain
a city name and a comma ie "Brooklyn, NY" or a zipcode}

\item{return_message}{if \code{TRUE} returns a message}

\item{...}{extra parameters}
}
\value{
a \code{tibble}
}
\description{
Acquires meta level information
for specified locations as of the
most recent time period
}
\examples{
vitality(locations = c("La Jolla, CA", "Manhattan, NY", "Bethany, DE", "10016"))
}
